Restructure elements into components
